feat: 30日講座メール配信準備
- MailTemplateSeeder: 30日分のメール内容を一括インポート
- Day 14.5対応: day_numberをdecimal型に変更
- SendSpecialEmailコマンド: Day 14.5などの特別メール配信用
  - current_dayを進めずに送信
  - 対象日指定可能（--target-day）

+++ backend/app/Models/MailTemplate.php
@@ -20,7 +20,7 @@ class MailTemplate extends Model
     protected function casts(): array
     {
         return [
-            'day_number' => 'integer',
+            'day_number' => 'float',
             'is_active' => 'boolean',
         ];
     }
